In template builder:
On the option tab.

a. The Select All button does select all items in the list. It checks all the boxes.

b. The list of reference lists is displaying the reference long and short names
for the reference lists.

c. Columns code, name, and column name are hidden columns, not shown to the user in the list.