Changed implementation of what is mandatory.
Added table DATAELEMENTMANDATORY. A data element is mandatory for a storage option, in the compilation year, if the data element is listed in this table. All data elements in non-compilation years are not mandatory.

Added tableDATAELEMENTMANDATORYSTATE. A data element is mandatory for a storage option, in the compilation year, in the state, if the data element is listed in this table. All data elements in non-compilation years are not mandatory. In 2000 which data elements are mandatory was dependent upon the state.